pyproject.toml
[tool.poetry]
name = "discord-template"
version = "0.1.0"
description = ""
authors = ["Selix <asel1x@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlmodel = "^0.0.22"
discord-py = "^2.4.0"
redis = "^5.1.1"
loguru = "^0.7.2"
alembic = "1.13.2"
dependency-injector = "^4.42.0"
asyncpg = "^0.29.0"
greenlet = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
